requirements.txt
python-dotenv~=1.0.1
Flask~=3.1.0
Flask-SocketIO~=5.5.1
openai~=0.28.0
requests~=2.32.3